[MERGE] Merge main to WH branch, and refactor
