+++ app/src/main/java/org/sopt/zepeto/data/response/BaseResponse.kt
@@ -1,6 +1,7 @@
 package org.sopt.zepeto.data.response
 
-data class BaseResponse<T>(
+data class
+BaseResponse<T>(
     val status: Int,
     val message: String,
     val data: T? = null
