+++ app/src/main/java/org/sopt/zepeto/data/service/ZepetoService.kt
@@ -3,9 +3,7 @@ package org.sopt.zepeto.data.service
 import org.sopt.zepeto.data.response.BaseResponse
 import org.sopt.zepeto.data.response.ResponseImages
 import retrofit2.Call
-import retrofit2.http.Body
 import retrofit2.http.GET
-import retrofit2.http.POST
 
 interface ZepetoService {
     @GET("images")
