+++ app/src/main/java/org/sopt/zepeto/ui/main/feed/FeedFragment.kt
@@ -24,7 +24,7 @@ class FeedFragment : Fragment() {
         _binding = null
     }
 
-    private fun initViewClickListener(){
+    private fun initViewClickListener() {
         binding.root.setOnClickListener {
             binding.clFeed.visibility = View.VISIBLE
         }
