+++ app/src/main/java/org/sopt/zepeto/ui/main/home/HomeFragment.kt
@@ -1,10 +1,10 @@
 package org.sopt.zepeto.ui.main.home
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import org.sopt.zepeto.databinding.FragmentHomeBinding
 
 class HomeFragment : Fragment() {
