+++ app/src/main/java/org/sopt/zepeto/ui/main/profile/ProfileFragment.kt
@@ -1,13 +1,13 @@
 package org.sopt.zepeto.ui.main.profile
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import org.sopt.zepeto.databinding.FragmentProfileBinding
 
-class ProfileFragment: Fragment() {
+class ProfileFragment : Fragment() {
     private var _binding: FragmentProfileBinding? = null
     private val binding get() = _binding ?: error("Binding이 초기화 되지 않았습니다")
