+++ app/src/main/java/org/sopt/zepeto/ui/main/world/WorldFragment.kt
@@ -1,13 +1,13 @@
 package org.sopt.zepeto.ui.main.world
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import org.sopt.zepeto.databinding.FragmentWorldBinding
 
-class WorldFragment: Fragment() {
+class WorldFragment : Fragment() {
     private var _binding: FragmentWorldBinding? = null
     private val binding get() = _binding ?: error("Binding이 초기화 되지 않았습니다")
